qcacld-2.0: Convert beacon data rate to hw code
User configures the beacon data rate in multiples of 100 Kbps
and FW expects the data rate in the form of hw rate codes.
So convert the data rates to hw rate code before configuring it to
firmware.

CRs-Fixed: 1020078

+++ CORE/MAC/inc/sirApi.h
@@ -694,7 +694,7 @@ typedef struct sSirSmeStartBssReq
     tANI_BOOLEAN            obssEnabled;
     uint8_t                 sap_dot11mc;
     bool                    vendor_vht_for_24ghz_sap;
-    uint8_t                 beacon_tx_rate;
+    uint16_t                beacon_tx_rate;
 } tSirSmeStartBssReq, *tpSirSmeStartBssReq;
 
 #define GET_IE_LEN_IN_BSS(lenInBss) ( lenInBss + sizeof(lenInBss) - \
