+++ CORE/SME/inc/csrInternal.h
@@ -387,8 +387,7 @@ typedef struct tagCsrRoamStartBssParams
 
     tSirAddIeParams     addIeParams;
     uint8_t             sap_dot11mc;
-    uint8_t             beacon_tx_rate;
-
+    uint16_t            beacon_tx_rate;
 }tCsrRoamStartBssParams;
